+++ keystorage/interlock/src/main/java/tech/pegasys/signers/interlock/InterlockClientFactory.java
@@ -21,49 +21,34 @@ import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientOptions;
 
-public class VertxHttpClientFactory {
-  private final Vertx vertx;
-  private final String host;
-  private final int port;
-  private final Path serverWhitelist;
+public class InterlockClientFactory {
 
   /**
-   * Constructor with parameters to create vertx HttpClient
+   * Create InterlockClient implementation using Vertx HttpClient
    *
    * @param vertx An instance of Vertx which will be used to create HttpClient
    * @param host The host name of interlock. Usually 10.0.0.1
    * @param port The port of intrelock server. Usually 443
    * @param serverWhitelist The server whitelist file containing host+port
-   *     server_certificate_fingerprint. If it doesn't exist or is empty, it will be created and
+   *     server_certificate_fingerprint. If file doesn't exist or is empty, it will be created and
    *     populated with server's TLS certificate fingerprint on first use.
    */
-  public VertxHttpClientFactory(
+  public static InterlockClient create(
       final Vertx vertx, final String host, final int port, final Path serverWhitelist) {
     Preconditions.checkNotNull(vertx, "vertx cannot be null");
     Preconditions.checkNotNull(host, "host cannot be null");
     Preconditions.checkNotNull(serverWhitelist, "server whitelist path cannot be null");
 
-    this.vertx = vertx;
-    this.host = host;
-    this.port = port;
-    this.serverWhitelist = serverWhitelist;
-  }
-
-  /**
-   * Create HttpClient
-   *
-   * @return HttpClient using Tuweni Vertx Trust Options
-   */
-  public HttpClient create() {
     final HttpClientOptions httpClientOptions =
         new HttpClientOptions()
-            .setDefaultHost(this.host)
-            .setDefaultPort(this.port)
+            .setDefaultHost(host)
+            .setDefaultPort(port)
             .setTryUseCompression(true)
             .setConnectTimeout(10_000)
             .setSsl(true)
             .setTrustOptions(trustServerOnFirstUse(serverWhitelist.toAbsolutePath(), true));
 
-    return vertx.createHttpClient(httpClientOptions);
+    final HttpClient httpClient = vertx.createHttpClient(httpClientOptions);
+    return new InterlockClientImpl(httpClient);
   }
 }
