+++ keystorage/interlock/src/main/java/tech/pegasys/signers/interlock/handlers/FileSizeHandler.java
@@ -40,9 +40,4 @@ public class FileSizeHandler extends AbstractHandler<Long> {
     }
     return -1L;
   }
-
-  @Override
-  public String body() {
-    return new JsonObject().put("path", path.getParent().toString()).put("sha256", true).encode();
-  }
 }
